requirements.txt
# Meeting agent dependencies
msal>=1.28.0,<2
fastapi>=0.110.0,<1
uvicorn>=0.29.0,<1
google-cloud-aiplatform>=1.50.0,<2
google-generativeai>=0.5.0,<1
langchain-text-splitters>=0.3.0,<0.4
supabase>=2.0.0,<3
httpx>=0.27.0,<1
python-dotenv>=1.0.0,<2
pydantic>=2.0.0,<3
